+++ src/services/languageResolver.ts
@@ -1,9 +1,11 @@
-import { PathLike, promises as fs, Dirent } from 'fs';
-import { extname, join } from 'path';
+import * as vscode from 'vscode';
+import { extname } from 'path';
 import AppMapAgent from '../agent/appMapAgent';
-import GitProperties from '../telemetry/properties/versionControlGit';
 import AppMapAgentRuby from '../agent/appMapAgentRuby';
 import AppMapAgentDummy from '../agent/AppMapAgentDummy';
+import extensionSettings from '../configuration/extensionSettings';
+import backgroundJob from '../lib/backgroundJob';
+import GitProperties from '../telemetry/properties/versionControlGit';
 
 const LANGUAGES = [
   {
@@ -177,72 +179,69 @@ const LANGUAGE_EXTENSIONS = LANGUAGES.reduce((memo, lang) => {
 /**
  * Register new AppMap agent CLI interfaces here.
  */
-const LANGUAGE_AGENTS = [
+export const LANGUAGE_AGENTS = [
   new AppMapAgentRuby(),
-  new AppMapAgentDummy('java'),
-  new AppMapAgentDummy('python'),
+  new AppMapAgentDummy('java', 'Java (Spring)'),
+  new AppMapAgentDummy('javascript', 'JavaScript (Node & Express)'),
+  new AppMapAgentDummy('python', 'Python (Django or Flask)', extensionSettings.pythonEnabled()),
   new AppMapAgentDummy('unknown'),
-  new AppMapAgentDummy('javascript'),
 ].reduce((memo, agent) => {
   memo[agent.language] = agent;
   return memo;
 }, {});
 
+type LanguageStats = Record<string, number>;
+
+const LANGUAGE_CACHE: Record<string, LanguageStats> = {};
+const LANGUAGE_CACHE_EXPIRY = 60 * 1000;
+const LANGUAGE_DISTRIBUTION_CACHE: Record<string, LanguageStats> = {};
+
+function folderPath(folder: vscode.WorkspaceFolder | string): string {
+  return typeof folder === 'string' ? folder : folder.uri.fsPath;
+}
+
 /**
  * Provides language identification APIs.
  */
 export default class LanguageResolver {
-  private static readonly LANGUAGE_DISTRIBUTION_CACHE = {};
-  private static readonly LANGUAGE_CACHE = {};
+  static clearCache(): void {
+    Object.keys(LANGUAGE_DISTRIBUTION_CACHE).forEach(
+      (key) => delete LANGUAGE_DISTRIBUTION_CACHE[key]
+    );
+    Object.keys(LANGUAGE_CACHE).forEach((key) => delete LANGUAGE_CACHE[key]);
+  }
 
-  public static async getLanguageDistribution(
-    rootDirectory: PathLike
-  ): Promise<Record<string, number>> {
-    const cachedValue = this.LANGUAGE_DISTRIBUTION_CACHE[rootDirectory as string];
+  static async getLanguageDistribution(
+    folder: vscode.WorkspaceFolder | string
+  ): Promise<LanguageStats> {
+    const cachedValue = LANGUAGE_DISTRIBUTION_CACHE[folderPath(folder)];
     if (cachedValue) {
       return cachedValue;
     }
 
-    const directories = [rootDirectory];
-    const extensions: Record<string, number> = {};
+    const extensions: LanguageStats = {};
     const gitProperties = new GitProperties();
-    await gitProperties.initialize();
+    await gitProperties.initialize(folderPath(folder));
 
-    for (;;) {
-      const currentDirectory = directories.pop() as string;
-      if (!currentDirectory) {
-        break;
-      }
+    const searchPattern = new vscode.RelativePattern(folderPath(folder), '**');
 
-      let files: Dirent[] = [];
-      try {
-        files = await fs.readdir(currentDirectory, { withFileTypes: true });
-      } catch {
-        continue;
-      }
-
-      files.forEach((f) => {
-        const absPath = join(currentDirectory, f.name);
-
-        if (f.isDirectory()) {
-          if (f.name !== '.git' && !gitProperties.isIgnored(absPath)) {
-            directories.push(absPath);
+    // VSCode will already respect the user's ignore list. We can supplement that with the .gitignore files.
+    await vscode.workspace.findFiles(searchPattern).then((files) => {
+      files
+        .filter((file) => !gitProperties.isIgnored(file.fsPath))
+        .forEach((file) => {
+          const fileExtension = extname(file.fsPath);
+          if (fileExtension !== '' && LANGUAGE_EXTENSIONS[fileExtension]) {
+            extensions[fileExtension] = (extensions[fileExtension] || 0) + 1;
           }
-          return;
-        }
-
-        const fileExtension = extname(f.name);
-        if (fileExtension !== '' && LANGUAGE_EXTENSIONS[fileExtension]) {
-          extensions[fileExtension] = (extensions[fileExtension] || 0) + 1;
-        }
-      });
-    }
+        });
+    });
 
     const languages = Object.entries(extensions).reduce((memo, [ext, count]) => {
       const language = LANGUAGE_EXTENSIONS[ext];
       memo[language] = memo[language] + count || count;
       return memo;
-    }, {} as Record<string, number>);
+    }, {} as LanguageStats);
 
     const totalFiles = Object.values(languages).reduce((a, b) => a + b, 0);
     if (totalFiles > 0) {
@@ -251,7 +250,7 @@ export default class LanguageResolver {
       });
     }
 
-    this.LANGUAGE_DISTRIBUTION_CACHE[rootDirectory as string] = languages;
+    LANGUAGE_DISTRIBUTION_CACHE[folderPath(folder)] = languages;
 
     return languages;
   }
@@ -260,35 +259,43 @@ export default class LanguageResolver {
    * Retrieve the most frequently used language id for a given directory. The language returned must be supported (i.e.,
    * it must be registered in LANGUAGE_AGENTS). If the language is not supported, returns 'unknown'.
    */
-  private static async identifyLanguage(rootDirectory: PathLike): Promise<string> {
-    const languages = await this.getLanguageDistribution(rootDirectory);
-    const best = Object.entries(languages).sort((a, b) => b[1] - a[1])?.[0]?.[0];
-    if (LANGUAGE_AGENTS[best]) return best;
+  private static async identifyLanguage(folder: vscode.WorkspaceFolder | string): Promise<string> {
+    let languageStats = LANGUAGE_CACHE[folderPath(folder)];
+    if (!languageStats) {
+      languageStats = await backgroundJob<LanguageStats>(
+        `appmap.languageDistribution.${folderPath(folder)}`,
+        LanguageResolver.getLanguageDistribution.bind(null, folder),
+        250
+      );
+
+      console.log(
+        `[language-resolver] Detected languages ${JSON.stringify(
+          languageStats
+        )} in project ${folderPath(folder)}`
+      );
+
+      LANGUAGE_CACHE[folderPath(folder)] = languageStats;
+      setTimeout(() => delete LANGUAGE_CACHE[folderPath(folder)], LANGUAGE_CACHE_EXPIRY);
+    }
+
+    const best = Object.entries(languageStats).sort((a, b) => b[1] - a[1])?.[0]?.[0];
+    const agent = LANGUAGE_AGENTS[best];
+    if (agent && agent.enabled) return best;
     return UNKNOWN_LANGUAGE;
   }
 
   /**
-   * Recursively walk the given `rootDirectory` for known source code files and returns the best-guess SUPPORTED
-   * language for the full directory tree. Ignores files and directories that are Git ignored. Results are cached for
-   * each directory for the lifetime of the extension. If the most used language is not supported, returns 'unknown'.
+   * Recursively inspects a folder for known source code files and returns the best-guess SUPPORTED
+   * language for the full directory tree. Ignores files and directories that are Git-ignored.
+   *
+   * @returns unknown if the most used language is not supported
    */
-  public static async getLanguage(rootDirectory: PathLike): Promise<string> {
-    let language = this.LANGUAGE_CACHE[rootDirectory as string];
-
-    if (!language) {
-      language = await this.identifyLanguage(rootDirectory);
-      this.LANGUAGE_CACHE[rootDirectory as string] = language;
-    }
-
-    return language;
+  public static async getLanguage(folder: vscode.WorkspaceFolder | string): Promise<string> {
+    return await this.identifyLanguage(folder);
   }
 
-  public static getAgentForLanguage(language: string): AppMapAgent | undefined {
-    return LANGUAGE_AGENTS[language];
-  }
-
-  public static async getAgent(rootDirectory: PathLike): Promise<AppMapAgent | undefined> {
+  public static async getAgent(rootDirectory: vscode.WorkspaceFolder): Promise<AppMapAgent> {
     const language = await this.getLanguage(rootDirectory);
-    return this.getAgentForLanguage(language);
+    return LANGUAGE_AGENTS[language];
   }
 }