+++ test/integration/util.ts
@@ -3,7 +3,7 @@ import { exec } from 'child_process';
 import { join } from 'path';
 import assert from 'assert';
 import AppMapService from '../../src/appMapService';
-import { touch } from '../../src/util';
+import { touch } from '../../src/lib/touch';
 
 export const FixtureDir = join(__dirname, '../../../test/fixtures');
 export const ProjectRuby = join(__dirname, '../../../test/fixtures/workspaces/project-ruby');
@@ -149,8 +149,14 @@ export async function executeWorkspaceOSCommand(cmd: string, workspaceName: stri
 
 async function cleanWorkspace(): Promise<void> {
   async function cleanProject(project: string) {
-    await executeWorkspaceOSCommand(`git checkout HEAD .`, project);
-    await executeWorkspaceOSCommand(`git clean -fd .`, project);
+    const commands = [`git checkout HEAD .`, `git clean -fd .`];
+    for (const command of commands) {
+      try {
+        await executeWorkspaceOSCommand(command, project);
+      } catch (e) {
+        console.log(e);
+      }
+    }
   }
   for (const project of PROJECTS) {
     await cleanProject(project);